+++ beacon-chain/state/setters.go
@@ -99,7 +99,7 @@ func (b *BeaconState) UpdateBlockRootAtIndex(idx uint64, blockRoot [32]byte) err
 	if !b.HasInnerState() {
 		return ErrNilInnerState
 	}
-	if len(b.state.BlockRoots) <= int(idx) {
+	if uint64(len(b.state.BlockRoots)) <= idx {
 		return fmt.Errorf("invalid index provided %d", idx)
 	}
 
@@ -155,7 +155,7 @@ func (b *BeaconState) UpdateStateRootAtIndex(idx uint64, stateRoot [32]byte) err
 	if !b.HasInnerState() {
 		return ErrNilInnerState
 	}
-	if len(b.state.StateRoots) <= int(idx) {
+	if uint64(len(b.state.StateRoots)) <= idx {
 		return errors.Errorf("invalid index provided %d", idx)
 	}
 
@@ -339,7 +339,7 @@ func (b *BeaconState) UpdateValidatorAtIndex(idx uint64, val *ethpb.Validator) e
 	if !b.HasInnerState() {
 		return ErrNilInnerState
 	}
-	if len(b.state.Validators) <= int(idx) {
+	if uint64(len(b.state.Validators)) <= idx {
 		return errors.Errorf("invalid index provided %d", idx)
 	}
 
@@ -401,7 +401,7 @@ func (b *BeaconState) UpdateBalancesAtIndex(idx uint64, val uint64) error {
 	if !b.HasInnerState() {
 		return ErrNilInnerState
 	}
-	if len(b.state.Balances) <= int(idx) {
+	if uint64(len(b.state.Balances)) <= idx {
 		return errors.Errorf("invalid index provided %d", idx)
 	}
 
@@ -447,7 +447,7 @@ func (b *BeaconState) UpdateRandaoMixesAtIndex(idx uint64, val []byte) error {
 	if !b.HasInnerState() {
 		return ErrNilInnerState
 	}
-	if len(b.state.RandaoMixes) <= int(idx) {
+	if uint64(len(b.state.RandaoMixes)) <= idx {
 		return errors.Errorf("invalid index provided %d", idx)
 	}
 
@@ -499,7 +499,7 @@ func (b *BeaconState) UpdateSlashingsAtIndex(idx uint64, val uint64) error {
 	if !b.HasInnerState() {
 		return ErrNilInnerState
 	}
-	if len(b.state.Slashings) <= int(idx) {
+	if uint64(len(b.state.Slashings)) <= idx {
 		return errors.Errorf("invalid index provided %d", idx)
 	}
 	b.lock.RLock()
